+++ packages/alfa-promise/src/global.ts
@@ -1,0 +1,9 @@
+// This file defines exports from the global `Promise` constructor for internal
+// use only.
+
+/**
+ * @internal
+ */
+const Builtin = Promise;
+
+export { Builtin as Promise };